+++ apps/client/src/components/Topbar/index.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
 import { Box, Heading, HStack, useColorModeValue } from '@chakra-ui/react';
-import { CaretLeft, CaretRight } from 'phosphor-react';
-import Button from '../ui/Button';
-import { useNavigate } from 'react-router-dom';
 import Search from './Search';
 import { useStore } from '~store/store';
 import { useViewMode } from '~hooks/useViewMode';
@@ -10,49 +7,7 @@ import shallow from 'zustand/shallow';
 import MobileDrawer from '../Sidebar/MobileDrawer';
 import SortConfig from './SortConfig';
 import ViewModeConfig from './ViewModeConfig';
-
-// TODO: fix navigation, I don't necessarily want native navigation here. I want to be able to disable the
-// buttons if theres is no forward/backward history. If I am on /books/id/pages/page, and I click the back button on
-// that page to take me to /boos/id, I don't want the back button to go to /books/id/pages/page.
-function Navigation() {
-	// const _location = useLocation();
-	const navigate = useNavigate();
-
-	// function navigateForward() {}
-
-	// function navigateBackward() {}
-
-	return (
-		<HStack
-			// >:( this won't work, probably some annoying thing with parent stack
-			// ml={0}
-			style={{ margin: 0 }}
-			spacing={1}
-			alignItems="center"
-			display={{ base: 'none', md: 'flex' }}
-		>
-			<Button
-				variant="ghost"
-				p="0.5"
-				size="sm"
-				_hover={{ bg: useColorModeValue('gray.200', 'gray.750') }}
-				onClick={() => navigate(-1)}
-			>
-				<CaretLeft />
-			</Button>
-
-			<Button
-				variant="ghost"
-				p="0.5"
-				size="sm"
-				_hover={{ bg: useColorModeValue('gray.200', 'gray.750') }}
-				onClick={() => navigate(1)}
-			>
-				<CaretRight />
-			</Button>
-		</HStack>
-	);
-}
+import Navigation from './Navigation';
 
 // FIXME: this is not good AT ALL for mobile. It *looks* fine, but the navigation is gone, the
 // sort/view mode buttons are gone, the sort config is gone,and the search bar is meh. I need to
